bqml/bq_xgboost/requirements.txt
google-cloud-bigquery==1.25.0
google-cloud
gcsfs
google-auth
google-cloud-logging